@zardus 2017-01-14
Now that the massive refactor is merged, could you rebase this on current master, @otibsa?

@ltfish, once that's done, any concerns about merging this?