@otibsa 2017-01-14
Rebased on master